Pull upstream changes from #37.
This adjusts the code after the said PR refactored the project
structure.

Reference: https://github.com/scrapinghub/web-poet/pull/37
