+++ web_poet/page_inputs/meta.py
@@ -1,0 +1,8 @@
+class Meta(dict):
+    """Container class that could contain any arbitrary data to be passed into
+    a Page Object.
+
+    Note that this is simply a subclass of Python's ``dict``.
+    """
+
+    pass
